🚚 rename [#28] 오타 수정
